Add fileName config to optionally override the document title while rendering and downloading.
