+++ src/components/DocumentNav.tsx
@@ -14,13 +14,6 @@ export const DocumentNav: FC<{}> = () => {
 
   if (documents.length <= 1 || !currentDocument) return null;
 
-  let fileName = currentDocument.uri;
-
-  const splitURL = fileName.split("/");
-  if (splitURL.length) {
-    fileName = splitURL[splitURL.length - 1];
-  }
-
   return (
     <Container id="doc-nav">
       <p id="doc-nav-info">
